htable: relicense under LGPL
Various LGPL components depend on it, via ccan/likely.  ccan/likely
really only needs it when CCAN_LIKELY_DEBUG is set, but making it a
conditional dependency is a bit nasty if defining that changes the
license.

So this is the simplest fix.  I might relicense under PD or BSD later,
since the likely module should probably have an even more liberal
license.