Add `container_of_var_off' macro to container_of
There is are certain use-cases when it is necessary to know the offset
of the member in a structure's memory layout. One such use-case can be
seen in `ccan/list/list.h' in macros `list_for_each' and
`list_for_each_safe'. This commit implements said functionality with
`container_of_var_off' macro.